+++ src/css/style.css
@@ -486,101 +486,6 @@ input[type=checkbox] {
 .appear {
   width: 50%; }
 
-.photo-post {
-  max-height: 50vh;
-  max-width: 100%; }
-
-/*comment-view*/
-.post-container {
-  bottom: 7.5%;
-  transition: 1s;
-  overflow: hidden;
-  position: absolute;
-  background-color: #fff;
-  width: 100%;
-  height: 0%; }
-
-.show-element {
-  height: 85%; }
-
-.go-back {
-  box-sizing: border-box;
-  border-bottom: solid 2px #817f7f;
-  width: 100%;
-  height: 6%;
-  text-align: left; }
-  .go-back .fa-arrow-left {
-    margin: 1%;
-    font-size: 22px;
-    color: #6e6d6d; }
-
-.settings-section {
-  position: relative;
-  width: 100%;
-  height: 94%; }
-
-#post-form {
-  background-color: #fff;
-  height: 100%;
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  align-items: center; }
-
-#option-public {
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: space-evenly;
-  align-items: center;
-  max-height: 100%;
-  width: 100%; }
-  #option-public #visibility-select {
-    width: 20vh;
-    font-size: 2.5vh;
-    padding: 0.8vh; }
-
-#post-form > div {
-  width: 100%;
-  height: 40%;
-  display: flex;
-  flex-direction: row; }
-
-.margin-photo {
-  margin-top: 6px; }
-
-#post-content {
-  font-size: 15px;
-  background-color: #fff;
-  padding: 3%;
-  padding-top: 20px;
-  border: none;
-  height: 60%;
-  width: 90%;
-  resize: none;
-  /* border: solid red; */ }
-
-#upload-photo, #change-photo {
-  display: none; }
-
-.photo-icon {
-  /* border: solid red; */
-  position: absolute;
-  top: 82%;
-  padding: 1%; }
-
-.fa-photo-video {
-  font-size: 26px;
-  color: #444444; }
-
-.btn-submit {
-  position: absolute;
-  top: 88%;
-  width: 50%;
-  border-radius: 5vh;
-  padding: 4%;
-  color: white;
-  background: #325aa0; }
-
 @font-face {
   font-family: 'Sifonn';
   src: url("../font/SIFONN_PRO.otf"); }
@@ -709,6 +614,117 @@ li {
 
 .options-appear {
   width: 100px; }
+
+.post-container {
+  bottom: 7.5%;
+  transition: 1s;
+  overflow: hidden;
+  position: absolute;
+  background-color: #fff;
+  width: 100%;
+  height: 85%; }
+  @media (min-width: 800px) {
+    .post-container {
+      width: 50%;
+      height: 0%;
+      border-radius: 3%; } }
+
+.show-element {
+  height: 85%; }
+  @media (min-width: 800px) {
+    .show-element {
+      border: solid #525050; } }
+
+.go-back {
+  box-sizing: border-box;
+  border-bottom: solid 2px #817f7f;
+  width: 100%;
+  height: 6%;
+  text-align: left; }
+  .go-back .fa-arrow-left {
+    margin: 1%;
+    font-size: 22px;
+    color: #6e6d6d; }
+
+.settings-section {
+  position: relative;
+  width: 100%;
+  height: 94%; }
+
+#post-form {
+  background-color: #fff;
+  height: 100%;
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  align-items: center; }
+
+#option-public {
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: space-evenly;
+  align-items: center;
+  max-height: 100%;
+  width: 100%; }
+  #option-public #visibility-select {
+    width: 20vh;
+    font-size: 2.5vh;
+    padding: 0.8vh; }
+
+#post-form > div {
+  width: 100%;
+  height: 40%;
+  display: flex;
+  flex-direction: row; }
+
+.margin-photo {
+  margin-top: 6px; }
+
+#post-content {
+  font-size: 15px;
+  background-color: #fff;
+  padding: 3%;
+  padding-top: 20px;
+  border: none;
+  height: 60%;
+  width: 90%;
+  resize: none;
+  /* border: solid red; */ }
+
+#upload-photo, #change-photo {
+  display: none; }
+
+.photo-icon {
+  position: absolute;
+  top: 82%;
+  padding: 1%; }
+  @media (min-width: 800px) {
+    .photo-icon {
+      top: 78%; } }
+
+.fa-photo-video {
+  font-size: 26px;
+  color: #444444; }
+
+.btn-submit {
+  position: absolute;
+  top: 88%;
+  width: 50%;
+  border-radius: 5vh;
+  padding: 4%;
+  color: white;
+  background: #325aa0; }
+  @media (min-width: 800px) {
+    .btn-submit {
+      width: 25%;
+      padding: 2%; } }
+
+.photo-post {
+  max-height: 50vh;
+  max-width: 100%; }
+
+.preview {
+  border: solid red; }
 
 .profile-section {
   width: 100%;
