+++ src/firebase/crud.js
@@ -1,6 +1,7 @@
-export const deletePost = (id) => {
+export const deletePost = (id, userId) => {
   db.collection('posts').doc(id).delete().then(() => {
-    console.log('Document successfully deleted!');
+    // llamar a colecion de usuarios
+    console.log(userId);
   })
     .catch((error) => {
       console.error('Error removing document: ', error);
