+++ src/index.html
@@ -11,30 +11,30 @@
   <script src="https://www.gstatic.com/firebasejs/7.14.5/firebase-storage.js"></script>
   <title>BUNKER</title>
   <link rel="stylesheet" href="css/style.css">
-  <link rel="icon" type="image/png" href="images/icono.png"/>
+  <link rel="icon" type="image/png" href="images/icono.png" />
 </head>
 
 <body>
   <section id="actual-view"></section>
   <section id="pantallaModal"></section>
- 
-<script>
-  // Your web app's Firebase configuration
-  const firebaseConfig = {
-    apiKey: "AIzaSyDeygw3xwQJVQiMVnznSDqq9nnn5EDW8_w",
-    authDomain: "bunker-e836e.firebaseapp.com",
-    databaseURL: "https://bunker-e836e.firebaseio.com",
-    projectId: "bunker-e836e",
-    storageBucket: "bunker-e836e.appspot.com",
-    messagingSenderId: "910029505675",
-    appId: "1:910029505675:web:429344be4dd11f0d69408e"
-  };
-  // Initialize Firebase
-  firebase.initializeApp(firebaseConfig);
-  const auth = firebase.auth();
-  const db = firebase.firestore();
-  const storage = firebase.storage();
-</script> 
+
+  <script>
+    // Your web app's Firebase configuration
+    const firebaseConfig = {
+      apiKey: "AIzaSyDeygw3xwQJVQiMVnznSDqq9nnn5EDW8_w",
+      authDomain: "bunker-e836e.firebaseapp.com",
+      databaseURL: "https://bunker-e836e.firebaseio.com",
+      projectId: "bunker-e836e",
+      storageBucket: "bunker-e836e.appspot.com",
+      messagingSenderId: "910029505675",
+      appId: "1:910029505675:web:429344be4dd11f0d69408e"
+    };
+    // Initialize Firebase
+    firebase.initializeApp(firebaseConfig);
+    const auth = firebase.auth();
+    const db = firebase.firestore();
+    const storage = firebase.storage();
+  </script>
   <script type="module" src="index.js"></script>
   <script src="https://kit.fontawesome.com/e43bba631d.js"></script>
