+++ src/sass/parts/_homeView.scss
@@ -136,115 +136,3 @@
     background-color: #fff;
     @include size(100%, 0);
 }
-.show-element{
-    height: 85%;
-}
-.menu-container{
-    top: 7.5%;
-    position: fixed;
-    right: 0%;
-    background-color: #373839;
-    @include size(0, 85%);
-    overflow: hidden;
-    transition: 0.7s;
-}
-.appear{
-    width: 50%
-} 
-.photo-post{
-    max-height: 50vh;
-    max-width: 100%;
-}
-
-/*comment-view*/
-.post-container{
-    bottom: 7.5%;
-    transition: 1s;
-    overflow: hidden;
-    position: absolute;
-    background-color: #fff;
-    @include size(100%, 0%);
-}
-.show-element{
-    height: 85%;
-}
-.go-back{
-    box-sizing: border-box;
-    border-bottom: solid 2px rgb(129, 127, 127);
-    @include size(100%, 6%);
-    text-align: left;
-    .fa-arrow-left{
-        margin:1%;
-        font-size: 22px;
-        color: rgb(110, 109, 109);
-    }
-}
-.settings-section{
-    // border: solid red; 
-    position: relative;
-    width: 100%;
-    height: 94%;
-}
-#post-form{
-    background-color: #fff;
-    height: 100%;
-    width: 100%;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-}
-#option-public{
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: space-evenly;
-    align-items: center;
-    // border: solid red;
-    max-height: 100%;
-    width: 100%;
-    #visibility-select{
-    width: 20vh;
-    font-size: 2.5vh;
-    padding: 0.8vh;
-}
-}
-#post-form > div{
-    @include size(100%, 40%);
-    display: flex;
-    flex-direction: row;
-}
-.margin-photo{
-    margin-top: 6px;
-}
-#post-content{
-    font-size: 15px;
-    background-color: #fff;
-    padding: 3%;
-    padding-top: 20px;
-    border: none;
-    height: 60%;
-    width: 90%;
-    resize: none;
-    /* border: solid red; */
-}
-#upload-photo, #change-photo{
-    display: none;
-}
-.photo-icon{
-    /* border: solid red; */
-    position: absolute;
-    top: 82%;
-    padding: 1%;
-}
-.fa-photo-video{
-    font-size: 26px;
-    color: rgb(68, 68, 68);
-}
-.btn-submit{
-    position: absolute;
-    top: 88%;
-    width: 50%;
-    border-radius: 5vh;
-    padding: 4%;
-    color: white;
-    background: rgb(50, 90, 160);
-}
