+++ src/view/index.js
@@ -2,11 +2,17 @@ import login from './login-view.js';
 import signup from './signup-view.js';
 import home from './home-view.js';
 import profile from './profile-view.js';
+import postform from './post-content.js';
+import editProfile from './edit-profile.js';
+import themes from './theme-options.js';
 
 const components = {
   login,
   signup,
   home,
   profile,
+  postform,
+  editProfile,
+  themes,
 };
 export { components };
