+++ src/view/profile-view.js
@@ -1,33 +1,13 @@
-import { renderPost } from './home-view.js';
+import { renderPost } from '../firebase-controller/renderpost.js';
 
 export default () => {
-  const profileContainer = document.createElement('div');
-  profileContainer.className = 'main-profile';
-  const profileView = `
-      <div class="profile-section">
-        <div class="profile-photos">
-          <img class="cover-profile cover-desktop">
-          <img class="profile profile-main circle" src="./images/profile-img-woman.png">
-        </div>
-        <div class="profile-information">
-          <h3>Usuario de BUNKER</h3>
-          <h5>Description</h5>
-        </div>
-      </div>
-      <div class="social no-lateral">
-        <div class="share-section container">
-          <img class="profile circle" src="./images/profile-img-woman.png">
-          <button class="share">What's on your mind?</button>
-        </div>
-        <ul class="core-rail" id="my-posts"></ul>
-      </div>`;
-  profileContainer.innerHTML = profileView;
-  const myPosts = profileContainer.querySelector('#my-posts');
+  const profileContainer = document.createElement('ul');
+  profileContainer.id = 'my-posts';
   // FIRESTORE GET DATA, SHOW JUST USER POSTS IN PROFILE
   auth.onAuthStateChanged((user) => {
     if (user) {
       db.collection('users').doc(user.uid).onSnapshot((doc) => {
-        myPosts.innerHTML = '';
+        profileContainer.innerHTML = '';
         const postsIds = doc.data().posts;
         const postsIdsKeys = Object.keys(postsIds);
         const ids = postsIdsKeys.map(currentId => postsIds[currentId]);
@@ -35,8 +15,8 @@ export default () => {
           // eslint-disable-next-line max-len
           const docs = userPosts.docs.filter(postDoc => ids.some(userPostId => postDoc.id === userPostId));
           // passing an array of documents
-          renderPost(docs).forEach((li) => {
-            myPosts.appendChild(li);
+          renderPost(docs, user.uid).forEach((li) => {
+            profileContainer.appendChild(li);
           });
         });
       });
